+++ src/main/java/uk/gov/hmcts/reform/prl/models/complextypes/serviceofapplication/OrdersToServeSA.java
@@ -1,25 +1,25 @@
-package uk.gov.hmcts.reform.prl.models.complextypes.ServiceOfApplication;
+package uk.gov.hmcts.reform.prl.models.complextypes.serviceofapplication;
 
 import lombok.Builder;
 import lombok.Data;
-import uk.gov.hmcts.reform.prl.enums.ServiceOfApplication.AmendDischargedVariedEnum;
-import uk.gov.hmcts.reform.prl.enums.ServiceOfApplication.AppointmentOfGuardianEnum;
-import uk.gov.hmcts.reform.prl.enums.ServiceOfApplication.BlankOrderEnum;
-import uk.gov.hmcts.reform.prl.enums.ServiceOfApplication.BlankOrderOrDirectionsEnum;
-import uk.gov.hmcts.reform.prl.enums.ServiceOfApplication.BlankOrderOrDirectionsWithdrawEnum;
-import uk.gov.hmcts.reform.prl.enums.ServiceOfApplication.ChildArrangementsSpecificProhibitedOrderEnum;
-import uk.gov.hmcts.reform.prl.enums.ServiceOfApplication.GeneralFormUndertakingEnum;
-import uk.gov.hmcts.reform.prl.enums.ServiceOfApplication.NonMolestationEnum;
-import uk.gov.hmcts.reform.prl.enums.ServiceOfApplication.NoticeOfProceedingsEnum;
-import uk.gov.hmcts.reform.prl.enums.ServiceOfApplication.NoticeOfProceedingsNonPartiesEnum;
-import uk.gov.hmcts.reform.prl.enums.ServiceOfApplication.NoticeOfProceedingsPartiesEnum;
-import uk.gov.hmcts.reform.prl.enums.ServiceOfApplication.OccupationEnum;
-import uk.gov.hmcts.reform.prl.enums.ServiceOfApplication.OtherUploadAnOrderEnum;
-import uk.gov.hmcts.reform.prl.enums.ServiceOfApplication.ParentalResponsibilityEnum;
-import uk.gov.hmcts.reform.prl.enums.ServiceOfApplication.PowerOfArrestEnum;
-import uk.gov.hmcts.reform.prl.enums.ServiceOfApplication.SpecialGuardianShipEnum;
-import uk.gov.hmcts.reform.prl.enums.ServiceOfApplication.StandardDirectionsOrderEnum;
-import uk.gov.hmcts.reform.prl.enums.ServiceOfApplication.TransferOfCaseToAnotherCourtEnum;
+import uk.gov.hmcts.reform.prl.enums.serviceofapplication.AmendDischargedVariedEnum;
+import uk.gov.hmcts.reform.prl.enums.serviceofapplication.AppointmentOfGuardianEnum;
+import uk.gov.hmcts.reform.prl.enums.serviceofapplication.BlankOrderEnum;
+import uk.gov.hmcts.reform.prl.enums.serviceofapplication.BlankOrderOrDirectionsEnum;
+import uk.gov.hmcts.reform.prl.enums.serviceofapplication.BlankOrderOrDirectionsWithdrawEnum;
+import uk.gov.hmcts.reform.prl.enums.serviceofapplication.ChildArrangementsSpecificProhibitedOrderEnum;
+import uk.gov.hmcts.reform.prl.enums.serviceofapplication.GeneralFormUndertakingEnum;
+import uk.gov.hmcts.reform.prl.enums.serviceofapplication.NonMolestationEnum;
+import uk.gov.hmcts.reform.prl.enums.serviceofapplication.NoticeOfProceedingsEnum;
+import uk.gov.hmcts.reform.prl.enums.serviceofapplication.NoticeOfProceedingsNonPartiesEnum;
+import uk.gov.hmcts.reform.prl.enums.serviceofapplication.NoticeOfProceedingsPartiesEnum;
+import uk.gov.hmcts.reform.prl.enums.serviceofapplication.OccupationEnum;
+import uk.gov.hmcts.reform.prl.enums.serviceofapplication.OtherUploadAnOrderEnum;
+import uk.gov.hmcts.reform.prl.enums.serviceofapplication.ParentalResponsibilityEnum;
+import uk.gov.hmcts.reform.prl.enums.serviceofapplication.PowerOfArrestEnum;
+import uk.gov.hmcts.reform.prl.enums.serviceofapplication.SpecialGuardianShipEnum;
+import uk.gov.hmcts.reform.prl.enums.serviceofapplication.StandardDirectionsOrderEnum;
+import uk.gov.hmcts.reform.prl.enums.serviceofapplication.TransferOfCaseToAnotherCourtEnum;
 import uk.gov.hmcts.reform.prl.models.documents.Document;
 
 import java.util.List;
